requirements.txt
Flask==3.0.2
PyMySQL==1.1.0
python-dotenv==1.0.1
Flask-Bootstrap==3.3.7.1 